Refactoring extract components from cpe.go
